librescan/requirements.txt
aniso8601==3.0.0
bottle==0.12.9
cffi==1.8.3
chdkptp.py==0.1.4
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-SocketIO==2.9.3
greenlet==0.4.13
itsdangerous==0.24
Jinja2==2.8
jpegtran-cffi==0.5.2
lupa==1.6
MarkupSafe==1.0
pexpect==4.2.1
Pillow==5.0.0
polib==1.0.7
ptyprocess==0.5.2
pycparser==2.18
pyjade==4.0.0
python-engineio==2.0.4
python-socketio==1.9.0
pytz==2018.3
PyYAML==3.11
six==1.11.0
Werkzeug==0.13
